app/requirements.txt
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
numpy==2.0.2
timm==1.0.12
Pillow==11.0.0
scipy==1.13.1
pandas==2.2.3
h5py==3.12.1
scikit-learn==1.5.2
matplotlib==3.9.3
